--- a/ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/AbstractEngineForkchoiceUpdated.java
+++ b/ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/AbstractEngineForkchoiceUpdated.java
@@ -381,7 +381,6 @@ public abstract class AbstractEngineForkchoiceUpdated extends ExecutionEngineJso
     return RpcErrorType.INVALID_PAYLOAD_ATTRIBUTES;
   }
 
-
   private static final String logMessage = "FCU({}) | head: {} | safe: {} | finalized: {}";
 
   private void logAtInfoFCUCall(