## PR description
I've noticed when working on troubleshooting block processing performance that FCU log is very important. It shows when the block worldstate data is really persisted to disk, as besu persists to disk on FCU. When FCU call is missing, besu needs to rollforward the world state in memory to execute the block, and this is impacting block processing performance performance.
Enabling FCU logs shows for example that during lock step sync with lodestar, Besu receives 32 newPayload calls followed by 32 FCU calls.
The 32th new payload call performance in this case is impacted by the fact it needs to rollforward the the wordstate in memory to be able to execute this block. 
This is not an important use case, but shows the importance of having the FCU log to understand what's happening on new payload execution.

**Before this PR**

```
{"@timestamp":"2025-04-28T14:37:30,890","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,099  (72143.....833a3)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,894","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,100  (3ebbd.....ff149)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,897","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,101  (47366.....ca544)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,900","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,102  (d2e6c.....8ff52)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,903","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,103  (eb7f5.....a93fe)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,907","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,104  (53072.....e12db)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,910","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,105  (41cc0.....cd212)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,915","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,106  (ea29a.....afacb)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,918","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,107  (f273c.....2973f)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.001s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,921","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,108  (85407.....b9b8e)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.001s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,925","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,109  (18592.....8c413)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,928","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,110  (119cc.....4bdc5)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,931","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,111  (8dd67.....c3c19)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,934","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,112  (08cea.....c92bf)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,937","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,113  (c9d1f.....1cabf)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,940","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,114  (040cf.....ba997)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.001s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,943","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,115  (7fb0b.....80ce5)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.001s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:30,951","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: b6c58.....5caf9 | finalized: 0a2dd.....d86ce | safeBlockHash: ed241.....f27be","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,218","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,116  (17a22.....7193c)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,221","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,117  (55121.....12015)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.001s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,224","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,118  (9ef72.....99c51)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.001s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,227","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,119  (b079b.....42025)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,230","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,120  (ce00d.....5f7b6)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.001s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,233","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,121  (940dc.....9678b)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,236","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,122  (eb4ad.....45d42)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,238","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,123  (7bb80.....bca23)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.001s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,241","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,124  (9e40d.....81ada)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.001s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,245","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,125  (21497.....df3e2)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,248","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,126  (1d608.....a5ca4)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
{"@timestamp":"2025-04-28T14:37:31,251","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #23,127  (16b6d.....11666)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.002s| mgas/s   0.00| peers:  7","throwable":""}
```

**After (with this PR)**

```
{"@timestamp":"2025-04-29T08:42:19,880","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,739  (90e0a.....943c8)|  514 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  45,895,767 (  1.1%)| exec time 0.263s| mgas/s 174.51| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:19,911","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,740  (d6d0e.....e5150)|    0 tx| 16 ws| 0 blobs| base fee      7 wei| gas used           0 (  0.0%)| exec time 0.028s| mgas/s   0.00| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:20,227","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,741  (c7262.....aac4e)|  584 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  52,999,122 (  1.3%)| exec time 0.309s| mgas/s 171.52| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:20,345","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,742  (d7972.....e4b7e)|  228 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  15,127,233 (  0.4%)| exec time 0.112s| mgas/s 135.06| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:21,370","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,743  (994e5.....2af2d)|   65 tx| 16 ws| 0 blobs| base fee      7 wei| gas used 130,005,785 (  3.3%)| exec time 1.021s| mgas/s 127.33| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:21,670","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,744  (1f162.....f66ce)|  564 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  51,897,982 (  1.3%)| exec time 0.293s| mgas/s 177.13| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:21,953","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,745  (87014.....0dd43)|  452 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  41,820,954 (  1.0%)| exec time 0.276s| mgas/s 151.53| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:22,173","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,746  (3f6c9.....1fa82)|  384 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  34,750,218 (  0.9%)| exec time 0.213s| mgas/s 163.15| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:22,335","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,747  (c2058.....96d5e)|  285 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  24,377,700 (  0.6%)| exec time 0.157s| mgas/s 155.27| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:22,570","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,748  (7a4fb.....4bba1)|  280 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  28,237,259 (  0.7%)| exec time 0.228s| mgas/s 123.85| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:22,719","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,749  (11076.....da8b8)|  275 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  21,955,600 (  0.5%)| exec time 0.142s| mgas/s 154.62| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:22,885","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,750  (7a516.....81e49)|  274 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  25,939,307 (  0.6%)| exec time 0.160s| mgas/s 162.12| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:23,099","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,751  (6f7e7.....9f2ff)|  280 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  28,154,962 (  0.7%)| exec time 0.209s| mgas/s 134.71| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:23,255","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,752  (66117.....9ff2f)|  279 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  24,278,630 (  0.6%)| exec time 0.150s| mgas/s 161.86| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:23,420","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,753  (79284.....6a7a1)|  277 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  26,045,987 (  0.7%)| exec time 0.160s| mgas/s 162.79| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:23,600","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,754  (9594d.....0eec5)|  272 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  25,780,364 (  0.6%)| exec time 0.175s| mgas/s 147.32| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:23,757","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,755  (8adc4.....b9c9c)|  279 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  24,297,840 (  0.6%)| exec time 0.152s| mgas/s 159.85| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:23,928","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,756  (404d6.....d680f)|  276 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  26,043,334 (  0.7%)| exec time 0.165s| mgas/s 157.84| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:24,280","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineNewPayload","message":"Imported #235,757  (52daa.....1533f)|  548 tx| 16 ws| 0 blobs| base fee      7 wei| gas used  49,965,378 (  1.2%)| exec time 0.345s| mgas/s 144.83| peers: 11","throwable":""}
{"@timestamp":"2025-04-29T08:42:24,327","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: 55c52.....379bd | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:24,445","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: dd55e.....f8078 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:24,509","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: c4a38.....31921 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:24,661","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: 773f5.....72696 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:24,829","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: 544d4.....1415b | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:24,917","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: 1e42f.....1f781 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:25,277","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: f6d91.....e9e93 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:25,393","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: 8eb87.....36ad6 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:25,445","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: 27c55.....0b489 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:25,487","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: 1b6c7.....f1136 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:25,571","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: 90e0a.....943c8 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
{"@timestamp":"2025-04-29T08:42:25,617","level":"INFO","thread":"vert.x-worker-thread-0","class":"AbstractEngineForkchoiceUpdated","message":"FCU(VALID) | head: d6d0e.....e5150 | finalized: eb921.....accb0 | safeBlockHash: 41ea7.....e56b7","throwable":""}
```


## Fixed Issue(s)
<!-- Please link to fixed issue(s) here using format: fixes #<issue number> -->
<!-- Example: "fixes #2" -->


### Thanks for sending a pull request! Have you done the following?

- [x] Checked out our [contribution guidelines](https://github.com/hyperledger/besu/blob/main/CONTRIBUTING.md)?
- [x] Considered documentation and added the `doc-change-required` label to this PR [if updates are required](https://wiki.hyperledger.org/display/BESU/Documentation).
- [x] Considered the changelog and included an [update if required](https://wiki.hyperledger.org/display/BESU/Changelog).
- [x] For database changes (e.g. KeyValueSegmentIdentifier) considered compatibility and performed forwards and backwards compatibility tests

### Locally, you can run these tests to catch failures early:

…